fix(debug): Added more debug capabilities
This commit includes an update to the Discord username on the 'SettingsUserPage' and introduces enhancements to debugging features in the code. The debug block has been made visible in the 'LyricsTile' class during debug mode, providing more detailed debug information such as the type of element being debugged and its size. A new Enum has been added to differentiate between 'Note' and 'Text' element types, improving the clarity of debug text. The change also fixes the method that returns the size of an item container in the 'NewLyricsScroller' class.