fix(debug): Minor changes to the debug logs
Modified the Write method in DebugHandler to include the exception's stack trace when logging errors. This will provide more information on the exception's origin and make debugging easier. Also added a limit to the number of log entries read from the file to the first 100, in order to prevent possible memory issues when dealing with large log files.