fix(lyrics): Just take it
A Debugger was included in the NewLyricsScroller to prevent application crashes caused by unhandled exceptions. The Debugger will log any exceptions that occur when rendering the scroller, with the goal of improving usability and stability. The debug logs were also added for specific actions to track and resolve possible future issues effectively