feat(lyrics): Completed TextOverlay for the new LyricsScroller
In this commit, multiple changes were made to provide a better visual experience for the user. The main change was in the Lyrics display UI. There were updates in the LyricsTile.axaml file, and methods for color interpolation were added in the ColorBrushExtension file. Improvements have been made to better handle the overlay of text and to provide better support for different selection modes. A new display mode 'SOLID' was added. Additionally, some codes were refactored, and unused ones were removed for optimization