feat(lyrics): The blur is back!
Added a new Blur functionality to the LyricsTile. This is to improve the user interface by providing a dynamic blur effect as lyrics transition changes. The BlurArea was added to the XAML and the corresponding Blur class was implemented in the code behind. Some adjustments were made in the LyricsScroller to properly check if lyrics are syncing. We ensured performance optimisation by directly checking if the blur effect is enabled and the size constraint is met before applying the effect. The result of this feature is a more visually appealing and readable lyrics scrolling experience