chore(lyrics): Minor changes
Introduced a new ScrollerElement class and refactored class LyricsScroller to
yield better calculation of lyrics display. The ScrollerElement holds a tuple of the index and size,
used in GetRenderedSize method, easing recall and manipulation of lyrics display properties.
The LyricsScroller was cleaned up to remove redundant code and improved syncing and resyncing functionalities.
The ScrollViewer transition easing property in LyricsScroller.axaml also got changed to CircularEaseOut for smoother scrolling effect