chore(lyrics): Scroller performance update
Simplified the LyricsScroller component by replacing redundant ItemsRepeater and hidden items used for size calculation with a straight-forward ItemsControl and a live-list of visual elements. The refactor led to the removal of unnecessary invisible controls, unneeded measurement cache, and made way for more optimized rendering process by directly binding to ViewModel's properties, reducing the time complexity and enhancing the performant behavior of the application. Advanced the LyricsScroller UX by making it more responsive and less prone to bugs with a more direct rendering of Lyrics tiles