chore(lyrics): Blur fixed, better performance
Added a new event LyricTimeChanged in LyricHandler which triggers upon change in lyrics' timestamp. This is to identify and distinguish between a simple change in lyrics and change in time of the lyrics. This new event is used in LyricsScroller to handle changes and ensure proper triggers. Furthermore, adjusted UI events associated with the lyrics display as they were causing misfires and incorrect functioning of the lyrics scroll. For instance, the functions associated with ScrollViewer have been refactored to be more efficient and predictable. Adjusted the code for UI elements involving lyrics for better alignment with the new event setup.