chore(lyrics): Fine-Tuning
The condition check for applying blur effect in LyricsTile.axaml.cs is simplified by removing unnecessary sub-condition `LyricsScroller.Instance.IsSynced`. This change avoids interfering with the blur functionality when the lyrics are not synced.

Additionally, redundant blank spaces in LyricsPage.axaml have been removed to keep the code clean. The margin of button inside the LyricsPage has been adjusted from "0,0,30,80" to "0,0,30,30" for better alignment and space utilization within the