chore(artwork): Updated image quality selection
Artwork fetching logic in DeezerCollector, SpotifyCollector, and MusixMatch, was updated to fetch and display smaller image sizes, enhancing performance and load times. Additionally, the ImageLoader.Source in LyricsPage.axaml was replaced with Image.Source to improve compatibility and reduce potential issues from the async image loader. Changes in LyricsPageViewModel.cs aim to directly assign the image from ArtworkFoundEventArgs, which condenses the code and eliminates the need for a separate artwork object.