chore(lyrics): Improved performance
Replaced SplitTextToLines method's return type from AList to ObservableCollection.
Removed "height" and "romanization" parameters from SplitTextToLines method, since they were not needed.
The usage of "height" in UpdateView method and _viewPort in InstanceOnEffectiveViewportChanged was modified.
As a result, the overhead of unnecessary computations is avoided, improving the efficiency of the text wrapping process.

This change makes the code cleaner, easier to read, and improves the performance of the lyrics text formatting process.