chore(codebase): Some improvements
AnimationUtils.cs was removed because it was no longer used in the project. Simplification of code and improvements in functionality made this utility class redundant. LyricsScroller.axaml.cs was updated to include a call to the ReSync() function immediately after filling visual elements. This ensures that lyrics are properly synchronized when updated.